requirements.txt
Django==1.6.5
django-braces==1.2.2
MySQL-python==1.2.5
nameko==1.8.1
pytz==2014.4
